% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS1.r
\name{Dulce_AnnotateMS1}
\alias{Dulce_AnnotateMS1}
\title{Annotate glycans in MS1 data}
\usage{
Dulce_AnnotateMS1(
  data,
  cwp = NULL,
  pdp = NULL,
  names = NULL,
  classes = "Unclassified",
  perfwhm = 0.5,
  cor_eic_th = 0.75,
  polarity = NULL,
  rtmin = 0,
  rtmax = Inf,
  pgp = NULL,
  ppm = NULL,
  mzabs = NULL,
  output = "all"
)
}
\arguments{
\item{data}{any class of object... Dulce message: Try me out baby.}

\item{output}{Character that determines what the function will return.
If \code{"main"}, a list with an \code{'XCMSnExp'} object (if possible) and the last achieved product will be returned.
If \code{"all"}, a list with all the partial products along the pipeline is returned.
If \code{"last"}, only the last achieved product will be returned.}
}
\value{
The last produced object before an error (or at least that is the idea)
}
\description{
AnnotateMS1 pipeline. The complete pipeline. Lots of things are gonna be assumed, many
defaults values would be taken... but hey! You will get something at the end.
Does it have any value? maybe, only time will tell...
...
And you, of course you have to tell if it has value or not (?). You are the one who should know!
Read some literature, don't be lazy.
}
\details{
Dobie dobie doo diroo diroo
}
\examples{
# Examples have to be made with a toy data object.

}
\seealso{

}
