% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS1.r
\name{Dulce_featuresAnnotate}
\alias{Dulce_featuresAnnotate}
\title{Annotate putative glycans}
\usage{
Dulce_featuresAnnotate(data, pgp = NULL, ppm = NULL, mzabs = NULL)
}
\arguments{
\item{data}{\code{data.frame} object with 'mzmin' and 'mzmax' columns.}

\item{pgp}{\code{predictGlycansParam()} object.}

\item{ppm}{M/z range for putative glycans in ppm units.}

\item{mzabs}{M/z range for putative glycans in absolute units. Changed to \code{NULL} if both ppm and mzabs are specified.}
}
\value{
data.frame
}
\description{
This functions is part of the AnnotateMS1 pipeline and
receives a \code{data.frame} object and merges it with the
output of \code{glycanPredict::predictGlycans()} function. \code{data} must have a
'mzmin' and a 'mzmax' columns.
}
\examples{
# Examples have to be made with a toy data object.

}
\seealso{
Dulce_AnnotateMS1
}
