% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS1.r
\name{Dulce_fetch}
\alias{Dulce_fetch}
\title{Fetch: find and group picks}
\usage{
Dulce_fetch(data, cwp = NULL, pdp = NULL, return_everything = F)
}
\arguments{
\item{data}{\code{XCMSnExp} object}

\item{cwp}{\code{CentWaveParam()} object. If not given, default \code{CentWaveParam()} values are used.}

\item{pdp}{\code{PeakDensityParam()} object. If not given, default \code{PeakDensityParam()} values are used.}

\item{return_everything}{Logical value. If FALSE, (default) a \code{XCMSnExp} object is returned.
If TRUE, a list, with (1) \code{XCMSnExp} object, (2) chromPeaks data.frame and (3) featureDefinitions data.frame, is returned.}
}
\description{
This functions is a robust pipeline that merges the
\code{findChromPeaks()} function and \code{groupChromPeaks()} function.
}
\details{
Dobie dobie doo diroo diroo
}
\examples{
# Examples have to be made with a toy data object.

}
\seealso{
xcms::CentWaveParam()
}
