% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS2.r
\name{Dulce_MS2CombNorm}
\alias{Dulce_MS2CombNorm}
\title{MS/MS Spectra Combination and Normalization}
\usage{
Dulce_MS2CombNorm(data, csp = NULL, norm.method = "max")
}
\arguments{
\item{data}{\code{MSpectra} object with msLevel=2 spectra.}

\item{csp}{\code{combineSpectraParam()} object.}

\item{`norm.method`}{normalization method for \code{MSnbase::normalise()} function.}
}
\value{
This function returns normalized and combined MS/MS spectra as \code{MSpectra} object.
}
\description{
This function cleans, combines and normalizes MS/MS spectra.
}
\seealso{

}
