% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS2.r
\name{Dulce_MS2PrecursorMatch}
\alias{Dulce_MS2PrecursorMatch}
\title{MS/MS Precursor Matching}
\usage{
Dulce_MS2PrecursorMatch(data, annotations)
}
\arguments{
\item{data}{\code{MSpectra} object.}

\item{annotations}{\code{data.frame} object with annotated MS features.}
}
\value{
This function returns a \code{data.frame} with annotated MS/MS spectra.
}
\description{
This function generates an annotated \code{data.frame} of the MS/MS
spectra information with the precursor annotations. It uses the
function \code{data.table::foverlaps()} on the mzmin and mzmax values of both
groups of spectra.
}
\seealso{

}
