% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS1.r
\name{Dulce_pickGroupPeaks}
\alias{Dulce_pickGroupPeaks}
\title{Pick and Group Peaks}
\usage{
Dulce_pickGroupPeaks(data, cwp = NULL, pdp = NULL, return_everything = F)
}
\arguments{
\item{data}{\code{XCMSnExp} object.}

\item{cwp}{\code{CentWaveParam()} object. If not given, default
\code{CentWaveParam()} values are used.}

\item{pdp}{\code{PeakDensityParam()} object. If not given, default
\code{PeakDensityParam()} values are used.}

\item{return_everything}{Logical value.}
}
\value{
If \code{return_everything} is set to FALSE (default), a \code{XCMSnExp} object is
returned. If TRUE, a list with (1) \code{XCMSnExp} object, (2) chromPeaks data.frame
and (3) featureDefinitions data.frame, is returned.
}
\description{
This function picks and groups peaks using the \code{findChromPeaks()}
and \code{groupChromPeaks()} functions from the \code{xcms} package.
}
\examples{
# Examples have to be made with a toy data object.

}
