% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS1.r
\name{Dulce_trimIsotopes}
\alias{Dulce_trimIsotopes}
\title{Trim Isotopes}
\usage{
Dulce_trimIsotopes(data, rtmin = 0, rtmax = Inf)
}
\arguments{
\item{data}{\code{xsAnnotate} object.}

\item{rtmin}{Minimum retention time value (in seconds) from which scans should be considered.}

\item{rtmax}{Maximum retention time value (in seconds) to which scans should be considered.}
}
\value{
This function returns a \code{data.frame} with collapsed information about the isotopic
signatures. A new column called "isotopes" is created, with a comma separated string
of the trimmed isotopes.
}
\description{
This functions receives a \code{xsAnnotate} object and trims its isotopes,
leaving only the rows with no isotopes, and collapsing those with isotopes into one row
(annotating those which were found in the "isotopes" column).
}
\examples{
# Examples have to be made with a toy data object.

}
\seealso{
Dulce_AnnotateMS1
}
