% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS1.r
\name{Dulce_AnnotateMS1}
\alias{Dulce_AnnotateMS1}
\title{Annotate glycans in MS1 data}
\usage{
Dulce_AnnotateMS1(
  data,
  cwp = NULL,
  pdp = NULL,
  names = NULL,
  classes = "Unclassified",
  perfwhm = 0.5,
  cor_eic_th = 0.75,
  polarity = NULL,
  rtmin = 0,
  rtmax = Inf,
  pgp = NULL,
  ppm = NULL,
  mzabs = NULL,
  output = "all"
)
}
\arguments{
\item{data}{Object of any of these classes: MSnExp, OnDiskMSnExp, XCMSnExp, xcmsSet,
xsAnnotate, data.frame.}

\item{cwp}{CentWaveParam() object. If not given, default CentWaveParam() values are used.}

\item{pdp}{PeakDensityParam() object. If not given, default PeakDensityParam() values are used.}

\item{names}{Character vector with length equals to the amount of samples. It defines
the identifier of each sample. If \code{NULL} (default), it generates names according
to: \code{sprintf('sample_\\\%03d', 1:nrow(data))} (sample_001, sample_002, etc).}

\item{classes}{Character vector with length equals to the amount of samples. It defines the main grouping of the samples. If \code{NULL} (default), it defines an "Unclassified" group for all samples.}

\item{perfwhm}{numeric value. It will be passed into \code{CAMERA::groupFWHM()} function.}

\item{cor_eic_th}{numeric value. It will be passed into \code{CAMERA::groupCorr()} function.}

\item{polarity}{character value, either "positive" or "negative". It will be passed into \code{CAMERA::findAdducts()} function.}

\item{rtmin}{Numeric value specifying the lower retention time boundary.}

\item{rtmax}{Numeric value specifying the higher retention time boundary.}

\item{pgp}{\code{predictGlycansParam()} object. If not given, default predictGlycansParam() values are used.}

\item{ppm}{M/z range for putative glycans in ppm units.}

\item{mzabs}{M/z range for putative glycans in absolute units. Changed to \code{NULL} if both, ppm and mzabs, are specified.}

\item{output}{Character object that determines what the function will return.
If \code{"main"}, a list with an \code{'XCMSnExp'} object (if possible) and the last achieved product will be returned.
If \code{"all"}, a list with all the partial products along the pipeline is returned.
If \code{"last"}, only the last achieved product will be returned.}
}
\value{
If no errors occurs during the execution of the function, then the output depends on th
\code{output} argument. If \code{all}, the function will return every partial product achieved. If \code{main},
XCMSnExp object (with peaks picked and grouped) and an annotated putative glycans \code{data.frame}  are
returned. If \code{last}, only the annotated putative glycans \code{data.frame} is returned. If an error occur
during the execution of the function, it saves the last partial product achieved, and tries to return
the specified output option.
}
\description{
This function constitutes the complete MS Data Annotation Pipeline. It is robust and smart
enough to \emph{understand} what kind of data is being inputed and continue the pipeline from the appropriate
step. It is constructed in such a way that when an error occurs, the last correctly executed product is
saved and returned.
}
\details{

}
\examples{
# Examples have to be made with a toy data object.

}
\seealso{

}
