% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS1.r
\name{Dulce_find}
\alias{Dulce_find}
\title{This function is the one I like the less.}
\usage{
Dulce_find(data, perfwhm = 0.5, cor_eic_th = 0.75, polarity = NULL)
}
\arguments{
\item{bla}{bla bla bla}
}
\value{

}
\description{
I will still use it in the main function, but major changes
have to be made.
}
\examples{
# Examples have to be made with a toy data object.

}
\seealso{
Dulce_AnnotateMS1
}
