% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS2.r
\name{Dulce_MS2CombNorm}
\alias{Dulce_MS2CombNorm}
\title{MS/MS Spectra Combination and Normalization}
\usage{
Dulce_MS2CombNorm(data, csp = NULL, norm.method = "max")
}
\description{
MS/MS Spectra Combination and Normalization
}
\seealso{

}
