% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMS2.r
\name{Dulce_MS2SpectraExtract}
\alias{Dulce_MS2SpectraExtract}
\title{MS/MS Spectra Extraction}
\usage{
Dulce_MS2SpectraExtract(data, features, annotations)
}
\description{
MS/MS Spectra Extraction
}
\seealso{

}
